Remove Data Tracker sync button and repurpose sync component to show signals
